requirements.txt
fastapi==0.112.1
uvicorn==0.27.1
sqlalchemy==2.0.32
pydantic==2.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
numpy==1.26.4
pandas==2.2.0
scikit-learn==1.4.0
tensorflow==2.16.2
joblib==1.3.2
alembic==1.13.1
pytest==8.0.1
httpx==0.26.0
email-validator==2.1.0
bcrypt==4.2.0
passlib==1.7.4

